Fixing core v2 installer as well as mic array

+++ installers/respeakerCoreV2.sh
@@ -3,15 +3,22 @@
 echo "############### Respeaker Core V2 installation ########################"
 echo "############## Please run this script with sudo #######################"
 
-USER=$(logname)
+if [[ $(logname) != "respeaker" ]]; then
+    echo "Cannot install respeaker Core V2 with another user than respeaker"
+    exit
+fi
 
 pip uninstall -y gpiozero
+pip --no-cache-dir install pydbus
+pip --no-cache-dir install pixel-ring
 
-cd /home/$USER
-curl https://raw.githubusercontent.com/respeaker/respeakerd/master/scripts/install_all.sh|bash
-cp -f ~/respeakerd/scripts/pixel_ring_server /usr/local/bin/
-chmod a+x /usr/local/bin/pixel_ring_server
-pixel_ring_server
+cd /home/respeaker
+
+rm /home/respeaker/install_all.sh
+wget https://raw.githubusercontent.com/respeaker/respeakerd/master/scripts/install_all.sh
+sudo su respeaker ./install_all.sh
+#cp -f /home/respeaker/respeakerd/scripts/pixel_ring_server /usr/local/bin/
+#chmod a+x /usr/local/bin/pixel_ring_server
 
 echo "############################## All done! ##############################"
 echo "############## Don't forget to turn on the SPI interface! #############"
